+++ CookManagement.VSA/Shared/DTOs/CountRequest.cs
@@ -5,6 +5,6 @@ namespace CookManagement.VSA.Shared.DTOs;
 public sealed record CountRequest
 {
     public string ProductCode { get; init; } = string.Empty;
-    public int Count { get; init; }
+    public Double Count { get; init; }
     public string? TimeZoneId { get; init; } = string.Empty;
 }
